remove oh-my-zsh
I was wondering if I really need oh my zsh, for quite a while I'm
customizing thing with a plugin like structure: env.d, and I was
wondering if except for the git prompt oh-my-zsh was really useful.

So with my buddy Claude, we tried to remove oh-my-zsh, prompt is now
using starship, and is supposed to be similar with the legacy prompt
(and even better with the upstream commit number difference being
displayed).

It is too soon to say that everything will be working as expected,
but I like the idea to bring it in house and have something a bit
more portable.

Now the next step will be to make it really portable, meaning we
should be able to use those dotfiles on mac or linux.